tiffany: Update for eleven and RevengeOS ify
Now it is ready for Xiaomi 5X, But time is limited and the problem has 
not been completely fixed. Now it has the following problems: 

1. Baseband version is unknown and can't read SIM card.
2. Keyboard light is not available.
3. You tell me.

Your pull request is always welcome, But I am in high school and may not 
be able to respond in time, thank you for your understanding.